requirements.txt
torch
torchvision
albumentations
tensorboard
pyyaml
scikit-learn
kaggle
matplotlib
seaborn